back/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
mysqlclient
django-cors-headers